bot-Hayati-main/requirements.txt
aiohttp==3.7.3
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
cachetools==4.2.1
certifi==2020.12.5
chardet==3.0.4
discord.py==1.6.0
google-api-core==1.25.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==5.1.0
oauthlib==3.1.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.6.0
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
wrapt==1.12.1
yarl==1.6.3
